Add support for pkg-config
DuckDB provides CMake package. It's useful to detect link options from
projects that use CMake, Meson and so on as a build system. But there
are build systems that don't support CMake package but support
pkg-config.

pkg-config support is useful for projects that use those build
systems.

This PR targets only libduckdb.so. libduckdb_static.a support is
out-of-scope of this PR.